requirements.txt
fastapi==0.115.0
uvicorn~=0.30.6
tortoise-orm==0.21.6
aerich==0.7.2
aiomysql==0.2.0
bcrypt~=4.2.0
PyJWT==2.9.0